Update README.md
Project brief:
Problem:
A small college library run by volunteers has donated its books to the case study library. They utilised a simple but effective spreadsheet-based cataloguing and loan system. Their books and data files are thus available for some experimental work. The library has a standard loan period of 14 days. A borrowed book returned after 19 days is 5 days late.
The text file books.csv contains a subset list of their books in CSV format. It has a header row. The CSV files are encoded as UTF-8. You may assume it is free from errors.
The text file bookloans.csv contains data on borrowed books in CSV format. It does not have a header row. Assume that this file has many errors. Each row holds a book_number, member_number, date_of_loan and date_of_return separated by commas.
Only transactions of books borrowed and returned during 2019 are valid. Many so-called loans are not for 2019.
• A date_of_return of 0 (zero) means that the book has not been returned yet.
• The file contains thousands of valid loan transactions but hundreds of invalid ones.
• Invalid loans include those that are for borrows of phantom books that are not in books.csv,
where the book has not yet been returned and where the book is borrowed or returned outside
of 2019.
• The date_of_loan is a single integer number representing the date of the loan in Microsoft Excel
Epoch Format. It represents the number of days elapsed since 1/1/1900. You may assume that 43464 was January 1st 2019 and 43819 was December 31st 2019. A date of borrow or return outside of that range would be an invalid date. It follows that a date_of_return of 0 is an invalid date
Requirements:
Provide a single Jupyter Notebook (. ipynb) file for the following tasks:
Tasks 1:
This task requires reading of books.csv and bookloans.csv into nested lists or dictionaries and ensuring you have a list of valid book loans. It requires processing to produce a printed report.
You need to consider validity of data used by your report. A report of books borrowed in 2019 is required listing the returned book number, book title, author, and the number of days borrowed.
This should be a valid list excluding the hundreds of false books and non-returned borrows. We want to show the least popular and most popular books.
We require the listing of books sorted by reverse order of the frequency that they were borrowed i.e., number of times someone borrowed the book.
The popularity is measured by how many times a book is borrowed, not for how many days it was borrowed. If there are two books and one is borrowed 27 times in 2019, and the other book is borrowed twice, the second book will be at the top of that list and deemed the least popular.
Write code to accomplish this.
  
Use functions where appropriate and comment your code often to show your assessor what you are doing and why.
Task 2:
The library is keen to know the interests of its readers to influence purchasing decisions.
The popularity is measured by how many times a book is borrowed, not for how many days it was borrowed.
Write code to produce a popularity report of all genres of books borrowed in 2019 and how many books are in that genre.
Write code to produce a popularity report of all sub-genres of books borrowed in 2019 and how many books are in that sub-genre.
Optionally the reports may be sorted.
You may use Markdown cells in the notebook to report on the quality of data.
Task 3:
The average duration that books were borrowed is of interest.
Write code to produce a report showing three averages.
Formatting to 2 decimal places is optional. For example,
print (f"On average a book is borrowed for {3.868645664:.2f} days") #Output: On average a book is borrowed for 3.87 days
For this task it would be useful to have a list for the period of days those books were borrowed and a list for the days late when a book is returned. The len() and sum() functions could be useful too. If a book is returned after 19 days, it is 5 days late.
The report should show:
a) The average number of days that a notional book was borrowed. Assume that if all the books were borrowed for 20,000 days in 2019 for a total number of 2,000 returns, the average number of days a book was borrowed was 10 days.
b) The percentage proportion of books returned late. Assume that if the total number of days that books were returned late was 2,000 days for books borrowed for 20,000 days, the proportion returned late was 10%.
c) The average late period. Assume that if books were returned late a total of 2,000 days over 2000 times that the average late period is 1 day. Write code including your own documented functions to accomplish the report.
Key points:
• Functions should be tested, and you should demonstrate how you have achieved this using cell markdown and comments of # or “””””” type.
• Cannot use panda
• Cannot use any statistical package
I will provide:
  
2 CSV files
books.csv (120 rows, 6 columns) bookloans.csv (2200 rows, 4 columns)